Reviewed by Rob.
Fixes: http://bugs.webkit.org/show_bug.cgi?id=11934

Adding/removing SVGPathSeg* items to a SVGPathElement (through SVGPathSegList
interface) has live update problems, as the notifyAttributeChange() function
isn't called anywhere. The context param is currently stored per SVGPathSeg,
which is not needed - as it can be done using SVGPathSegList completely. It
wastes memory, that is fixed. In order to get rid of the context param in
SVGPathSeg, we need a way to map SVGPathSeg* objects to SVGStyledElement*
objects ("context element"). That is done using a new HashMap in the
SVGDocumentExtensions class.


Identifier: 15442@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@18393 268f45cc-cd09-0410-ab3c-d52691b4dbfc