Fix two typos spotted by MSVC /W3
Both typos were a semicolon following the condition of an if. This
would cause the intended body of the if to always execute.

Unfortunately neither of these fixes are testable.
BackForwardList::goForward is only called from WebBackForwardList, and
the createDragImageFromImage bug would only manifest itself when
CreateCompatibleDC failed.

Reviewed by Mark.

* history/BackForwardList.cpp:
(WebCore::BackForwardList::goForward):
* platform/win/DragImageWin.cpp:
(WebCore::createDragImageFromImage):


Identifier: 18643@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@23875 268f45cc-cd09-0410-ab3c-d52691b4dbfc