Reviewed by Eric.
Fix bad abstraction in ImageBuffer class.
While the class itself is platform-aware and thus it's location in platform/graphics
is fine, it contains the "renderSubtreeToImage" method which operates on RenderObject.

As it's SVG specific I decided to move this method into SVGRenderSupport, to avoid
implicit platform/ <-> rendering/ dependencies.


Identifier: 20574@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@26595 268f45cc-cd09-0410-ab3c-d52691b4dbfc