Fix <rdar://problem/5365030> calling dataWithPDFInsideRect on an SVG with a gradient crashes (14780)
Reviewed by Anders.

When drawing directly to PDF CG may delay the use of the gradient function until outside our
standard drawing path, which in turn could let us invalidate the caches before they were used.

To work around this we now store the cached stops in a RefCounted object, so that we can ensure
that cache exists as long as required.


Identifier: 21702@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@27781 268f45cc-cd09-0410-ab3c-d52691b4dbfc