Reviewed by Eric.
Fixes: http://bugs.webkit.org/show_bug.cgi?id=16445 (Refactor EventTargetNode & JSEventTargetNode for an upcoming SVG patch)

Split up JSEventTargetNode in JSEventTargetNode & JSEventTargetBase - where most functionality has been moved down
in the base class. Applied the same refactorization to EventTargetNode.

This makes it possible for the upcoming patch implementing the EventTarget interface for SVGElementInstance
to share as much code as possible with the EventTargetNode classes.


Identifier: 22621@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@28756 268f45cc-cd09-0410-ab3c-d52691b4dbfc