EventHandler changes/fixes in QWebPage:     -mouse{Press,Move,Release}Event:         Send the event always to the mainFrame of the QWebPage.
    -contextMenuEvent, key{Press,Release}Event:
        Send the event to the focused frame.

This is following the Windows port and fixes a issue with the
Web Inspector where we were sending the events to a wrong frame.

It is guaranteed that the mainFrame will always have an eventHandler
and frameView set. There is no need to check for this in QWebPage.


Identifier: 23349@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@29517 268f45cc-cd09-0410-ab3c-d52691b4dbfc