Support smart copy and paste during drag and drop
Reviewed by Alice

In order for the drag and drop tests in Windows DRT to pass
we need to support smart cut and paste operations during
drag and drop on windows.

There is no layout test as drag and drop is still unsupported
on windows.  Once supported smart drag/drop is tested by existing
tests


Identifier: 23429@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@29597 268f45cc-cd09-0410-ab3c-d52691b4dbfc