Fix return type conversions from Qt slots to JS values.
This also fixes fast/dom/open-and-close-by-DOM.html, which called
layoutTestController.windowCount().

When constructing the QVariant that holds the return type we cannot
use the QVarian(Type) constuctor as that will create a null variant.
We have to use the QVariant(Type, void *) constructor instead, just
like in QMetaObject::read() for example.


Identifier: 23452@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@29620 268f45cc-cd09-0410-ab3c-d52691b4dbfc