Add fast multi-level scope lookup
Reviewed by Geoff, Darin and Weinig

Add logic and AST nodes to provide rapid variable resolution across
static scope boundaries.  This also adds logic that allows us to skip
any static scopes that do not contain the variable to be resolved.

This results in a ~2.5% speedup in SunSpider, and gives a 25-30% speedup
in some simple and ad hoc closure and global variable access tests.


Identifier: 24794@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@31114 268f45cc-cd09-0410-ab3c-d52691b4dbfc