Tor Arne Vestbø <tavestbo@trolltech.com>
Fixes: Fix strange button sizes (push, radio, and check) in QtWebKit.

We now use pixel metrics instead of fixed sizes for the
button margins, or if the style supports layout rects we
use the delta between the content rect and the layout rect.

In adition we compensate for the difference between the size
for content and the actual layout rect (if supported) by
reporting back the layout rect to WebKit for page layout,
and then inflating the rect we actually paint to.


Identifier: 25846@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@32559 268f45cc-cd09-0410-ab3c-d52691b4dbfc