<rdar://problem/5956612> ibtool crashes during build
SCDynamicStoreCreate will return null if it cannot create a dynamic store object.  We need
to check for failure rather than continuing with a null store as this results in us
passing a null CFRunLoopSourceRef to CFRunLoopAddSource, which promptly crashes.

Reviewed by Tim Hatcher.

* platform/network/mac/NetworkStateNotifierMac.cpp:
(WebCore::NetworkStateNotifier::NetworkStateNotifier): Null check SCDynamicStoreCreate and
SCDynamicStoreCreateRunLoopSource as it is possible for them to fail.


Identifier: 26559@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@34050 268f45cc-cd09-0410-ab3c-d52691b4dbfc