<rdar://problem/5951874> WebHTMLHighlighter mistakenly gained two new methods, causing compile warnings
Reviewed by Darin Adler.

_pauseNullEventsForAllNetscapePlugins and _resumeNullEventsForAllNetscapePlugins ended up being declared both in
WebHTMLViewInternal.h and as members of the WebHTMLHighlighter protocol in WebHTMLViewPrivate.h.  They don't belong
in the protocol, but they do need to be available outside of WebKit so they're being moved to the correct location
in WebHTMLViewPrivate.h and removed from WebHTMLViewInternal.h.

* WebView/WebHTMLView.mm:
(-[WebHTMLView _pauseNullEventsForAllNetscapePlugins]):
(-[WebHTMLView _resumeNullEventsForAllNetscapePlugins]):
* WebView/WebHTMLViewInternal.h:
* WebView/WebHTMLViewPrivate.h:


Identifier: 27056@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@34608 268f45cc-cd09-0410-ab3c-d52691b4dbfc