Bug 18774: SQUIRRELFISH: print meaningful error messages <https://bugs.webkit.org/show_bug.cgi?id=18774> <rdar://problem/5769353> SQUIRRELFISH: JavaScript error messages are missing informative text
Reviewed by Cameron Zwarich

Add support for decent error messages in JavaScript.  This patch achieves this by providing
ensuring the common errors and exceptions have messages that provide the text of expression
that trigger the exception.  In addition it attaches a number of properties to the exception
object detailing where in the source the expression came from.


Identifier: 27676@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@35245 268f45cc-cd09-0410-ab3c-d52691b4dbfc