Improve switch performance.
Reviewed by Geoff Garen and Sam Weinig.

Improve switch performance by converting to a hashmap based jump
table to avoid the sequence of dispatches that would otherwise be
needed.  This results in a 9-19x performance win for string switches
based on ad hoc testing, and a 6x improvement for integer switch
statements.  SunSpider reports a 1.2% progression.


Identifier: 27739@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@35309 268f45cc-cd09-0410-ab3c-d52691b4dbfc