<rdar://problem/6111648> JavaScript exceptions fail if the scope chain includes the global object
Reviewed by Alexey Proskuryakov.

In an attempt to remove the branch I just added to KJS::depth I
used the existence of a Variable Object at a point in the scope
chain as an indicator of function or global scope activation.
However this assumption results in incorrect behaviour if the
global object is injected into the scope chain with 'with'.


Identifier: 27869@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@35445 268f45cc-cd09-0410-ab3c-d52691b4dbfc