Bug 19359: JavaScriptCore behaves differently from FF2/3 and IE when handling context in catch statement <https://bugs.webkit.org/show_bug.cgi?id=19359>
Reviewed by Cameron Zwarich

Make our catch behave like Firefox and IE, we do this by using a StaticScopeObject
instead of a generic JSObject for the scope node.  We still don't make use of the
fact that we have a static scope inside the catch block, so the internal performance
of the catch block is not improved, even though technically it would be possible to
do so.


Identifier: 27951@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@35533 268f45cc-cd09-0410-ab3c-d52691b4dbfc