Add an ASCII fast-path to toLowerCase and toUpperCase.
The fast path speeds up the common case of an ASCII-only string by up to 60% while adding a less than 5% penalty
to the less common non-ASCII case.

This also removes stringProtoFuncToLocaleLowerCase and stringProtoFuncToLocaleUpperCase, which were identical
to the non-locale variants of the functions.  toLocaleLowerCase and toLocaleUpperCase now use the non-locale
variants of the functions directly.

Reviewed by Sam Weinig.

* kjs/StringPrototype.cpp:
(JSC::stringProtoFuncToLowerCase):
(JSC::stringProtoFuncToUpperCase):

Identifier: 29317@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@37049 268f45cc-cd09-0410-ab3c-d52691b4dbfc