Reviewed by Adam Roben.
Fixes: https://bugs.webkit.org/show_bug.cgi?id=23433

Add InputElement abstraction, similar to the FormControlElement refactorization
done a month ago. The goal is to share as much code as possible between HTML/WMLInputElement.
In a previous patch the RenderTextControl class has been split-up in RenderTextControlSingle/MultiLine,
RenderTextControl itself only depends on FormControlElement, RenderTextControlSingleLine still
depends on HTMLInputElement directly -> change that and make it use the new InputElement abstraction.

A lot of code from HTMLInputElement now lives in InputElement, as static member functions - the
InputElement class itself is an abstract virtual class, just like ScriptElement. HTML/WMLInputElement
derive from InputElement, and hold a InputElementData member variable, that they pass to the
static functions in InputElement. The abstraction is equal to the one chosen for HTML/SVGScriptElement.


Identifier: 32261@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@40065 268f45cc-cd09-0410-ab3c-d52691b4dbfc