Move Frame::isFrameSet to Document::isFrameSet
Changed the FrameLoader callsite. I assume that as this
point m_frame.document() might return a null pointer. Change
the condition to either not having a document or not having a
frame set. This should be semantically equivalent to the
old code.

Make Document::body() const to be able to call it from within
Document::isFrameSet. Leave the isHTMLDocument() check as this
is a stronger condition than simply having a HTMLElement in the
DOM.

Identifier: 32626@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@40443 268f45cc-cd09-0410-ab3c-d52691b4dbfc