<rdar://problem/6050421> JavaScript register file should remap to release physical pages accumulated during deep recursion
Reviewed by Geoff Garen

We now track the maximum extent of the RegisterFile, and when we reach the final
return from JS (so the stack portion of the registerfile becomes empty) we see
if that extent is greater than maxExcessCapacity.  If it is we use madvise or
VirtualFree to release the physical pages that were backing the excess.

Identifier: 34892@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@42842 268f45cc-cd09-0410-ab3c-d52691b4dbfc