Fix <https://bugs.webkit.org/show_bug.cgi?id=25332>. Bug 25332: Plug-in inserted in to DOM of PluginDocument fails to load
Reviewed by Darin Adler.

Only the first plug-in in a PluginDocument can use the main resource data.  For all
subsequent plug-ins, such as the plug-in that ClickToFlash inserts when swapping itself
out for the real Flash plug-in, we need to load the resource data in the same manner
as for other embedded plug-ins.

* loader/FrameLoader.cpp:
(WebCore::FrameLoader::loadPlugin):

Identifier: 35213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43181 268f45cc-cd09-0410-ab3c-d52691b4dbfc