Fix crash due to incorrectly using an invalid scopechain
Reviewed by Gavin Barraclough.

stringProtoFuncReplace was checking for an exception on a CachedCall
by asking for the cached callframes exception.  Unfortunately this
could crash in certain circumstances as CachedCall does not guarantee
a valid callframe following a call.  Even more unfortunately the check
was entirely unnecessary as there is only a single exception slot per
global data, so it was already checked via the initial exec->hadException()
check.

To make bugs like this more obvious, i've added a debug only destructor
to ScopeChainNode that 0's all of its fields.  This exposed a crash in
the standard javascriptcore tests.


Identifier: 35364@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43341 268f45cc-cd09-0410-ab3c-d52691b4dbfc