Improve function call forwarding performance
Reviewed by Gavin Barraclough

Make creation of the Arguments object occur lazily, so it
is not necessarily created for every function that references
it.  Then add logic to Function.apply to allow it to avoid
allocating the Arguments object at all.  Helps a lot with
the function forwarding/binding logic in jQuery, Prototype,
and numerous other JS libraries.


Identifier: 35538@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43559 268f45cc-cd09-0410-ab3c-d52691b4dbfc