[GTK+] Make TestNescpaePlugin loadable in mozilla
https://bugs.webkit.org/show_bug.cgi?id=27345

Currently the NPNetscapeFuncs of WebKit contains two more
entries than the gecko version. This means comparing the
aMozillaVTable->size with the size of NPNetscpaeFuncs will
fail and the plugin will not be loaded.

For testing it is nice to be able to load the plugin into
gecko and currently we are not using the two extra methods.
Remove the test to be able to use the plugin in gecko.

Identifier: 37660@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@46036 268f45cc-cd09-0410-ab3c-d52691b4dbfc