Bug 27923: run-webkit-tests should not invoke "ps" repeatedly on Mac OS X Fix <https://bugs.webkit.org/show_bug.cgi?id=27923>.
Reviewed by Sam Weinig and David D. Kilzer.

run-webkit-tests invokes "ps" after each test on Mac OS X to determine whether DumpRenderTree is
in the process of crashing.  We can do the same test more efficiently with a simple call to sysctl.
I couldn't find any built-in way of doing this directly from perl, so we do it from a simple C module
instead.  This speeds up run-webkit-tests by around 8%.

* DumpRenderTree/DumpRenderTree.xcodeproj/project.pbxproj:
* DumpRenderTree/mac/PerlSupport/DumpRenderTreeSupport.c: Added.
(processIsCrashing):
* DumpRenderTree/mac/PerlSupport/DumpRenderTreeSupportTiger.pm: Added.  Tiger does not have the SWIG tool
that is used to generate the Perl binding to the C code, so we fall back to the old implementation on Tiger.
* DumpRenderTree/mac/PerlSupport/Makefile: Added.
* Scripts/run-webkit-tests:

Identifier: 38288@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@46682 268f45cc-cd09-0410-ab3c-d52691b4dbfc